Fix implicit overflow in AGAST
	modified:   Agast5_8.h
	modified:   Agast7_12d.h
	modified:   Agast7_12s.h
	modified:   OAST9_16.h

+++ cvt/vision/features/agast/Agast5_8.h
@@ -38,8 +38,8 @@ namespace cvt
             void detect( const Image& img, uint8_t threshold, FeatureSetWrapper& features, size_t border );
 
         private:
-            size_t  _lastStride;
-            int16_t offset0, offset1, offset2, offset3, offset4, offset5, offset6, offset7;
+            ssize_t _lastStride;
+            ssize_t offset0, offset1, offset2, offset3, offset4, offset5, offset6, offset7;
 
             int cornerScore( const uint8_t* p, uint8_t thresh ) const;
